requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
jose
passlib
alembic
python-dotenv
pytest
fastapi-testclient
